requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
Django==4.1.3
django-celery-beat==2.4.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
flake8==6.0.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
render==1.0.0
signals==0.0.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
vine==5.0.0
virtualenv==20.16.7
wcwidth==0.2.5
